requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
PyJWT # For authentication